switch pof to decimial

+++ user_files/views.py
@@ -3,11 +3,10 @@ from ajaxuploader.views import AjaxFileUploader
 from django.contrib.auth.decorators import login_required
 from django.contrib.admin.views.decorators import staff_member_required
 import os, tempfile, zipfile
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse
 from django.core.servers.basehttp import FileWrapper
 import mimetypes
 from .models import UploadedFile
-
 
 @staff_member_required
 def send_file(request, id):
@@ -35,6 +34,21 @@ def delete_uploaded_file(request):
     file_id = request.POST.get('file_id', None)
     selected_file = get_object_or_404(
         UploadedFile, id=file_id, user=request.user)
-    data = {"name": selected_file.supporting_document.name, "id": selected_file.id}
-    selected_file.delete()
+
+    protected = False
+    message = 'File deleted.'
+    if selected_file.application is not None:
+        if selected_file.application.frozen == True:
+            protected = True
+            message = 'The application this file is linked to is frozen for editing and so the file can not be deleted. Email chris.hartley@renewindianapolis.org for help.'
+
+    data = {
+        "name": selected_file.supporting_document.name,
+        "id": selected_file.id,
+        "protected": protected,
+        "message": message
+        }
+
+    if protected == False:
+        selected_file.delete()
     return JsonResponse(data)
